Fix read checkbox and add title checker
Toggling the read checkbox will now successfully change the status
of the 'read' property on the book object to true or false. Also
addded a title checker that will prevent the user from inserting
a new book if there's already an existing book in the library
with the same title.